Add menu configuration option to swap Blue and Green colors.
The HUB75 64x64 display that I purchased from Aliexpress had the B & G colors swapped. This seems to be a common problem that people are facing.

This code change adds code that will swap the B & G pins, as well as a menu configuration for esp-idf that makes it easy to enable/disable.